netracare_backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-RESTX
bcrypt
PyJWT
